+++ TreasureHuntBackend/bin/src/main/java/com/example/demo/models/TreasureHunt.class
