+++ TreasureHuntBackend/bin/src/main/java/com/example/demo/models/customObject.class
