--- demo/src/main/resources/META-INF/persistence.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
- 
-<persistence version="2.0"
-    xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
-    
-    <!-- Define a name used to get an entity manager. Define that you will 
-    complete transactions with the DB  -->
-    <persistence-unit name="jpatest" transaction-type="RESOURCE_LOCAL">
-    
-        <!-- Define the class for Hibernate which implements JPA -->
-        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <!-- Define the object that should be persisted in the database -->
-        <class>com.example.demo.models.TreasureHuntStep</class>
-        <properties>
-            <property name="javax.persistence.jdbc.driver" value="com.mysql.jdbc.Driver" />
-            <property name="javax.persistence.jdbc.url" value="jdbc:mysql://localhost/treasurehunt" /> 
-            <property name="javax.persistence.jdbc.user" value="root" />
-            <property name="javax.persistence.jdbc.password" value="" />
-        </properties>
-    </persistence-unit>
-</persistence>
